refactored mutate_rows logic into helper function
